requirements.txt
evdev==1.6.1
kink==0.7.0
numpy==1.26.0
Pillow==10.0.1
pynput==1.7.6
PyOpenGL==3.1.7
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
pyqtgraph==0.13.3
PySide6==6.5.3
PySide6-Addons==6.5.3
PySide6-Essentials==6.5.3
python-xlib==0.33
shiboken6==6.5.3
six==1.16.0
typing_extensions==4.8.0
